Catalog

Many common or useful matroids.

Matroids.Catalog.Block_10_5

Matroids.Catalog.Fano

Matroids.Catalog.NonFano

Matroids.Catalog.Pappus

Matroids.Catalog.Q6

Matroids.Catalog.R6

Matroids.Catalog.Uniform

Functions

Matroids.Catalog.Block_10_5 — Method

Block_10_5()

Return the paving matroid whose non-spanning circuits form of a 3-(10, 5, 3) design, represented as circuit closures.

Matroids.Catalog.Fano — Method

Fano()

Returns the Fano matroid, represented as a linear matroid over GF(2).

Matroids.Catalog.NonFano — Method

NonFano()

Returns the non-Fano matroid, represented as a linear matroid over GF(3).

Matroids.Catalog.Pappus — Method

Pappus()

Returns the Pappus matroid, represented as circuit closures.

Matroids.Catalog.Q6 — Method

Q6()

Returns the Q6 matroid, represented as a linear matroid over GF(4).

Matroids.Catalog.R6 — Method

R6()

Returns the R6 matroid, represented as a linear matroid over GF(3).

Matroids.Catalog.Uniform — Method

Uniform(r, n)

Return the uniform matroid of rank r on n elements represented as circuit closures.