List of Functions and their Documentation Messages

Graphs.LinAlg.contract — Method

contract(M::Matroid, X)

Form a new matroid by contracting the elements of X in the matroid M. Here, X may be a Set or a Vector of integer values, or a single integer.

Graphs.LinAlg.incidence_matrix — Method

incidence_matrix(g::EasyMultiGraph)

Create a signed incidence matrix for g. Multiple edges become repeated columns. Loops become all-zero columns. The order of the columns exactly matches the output of edges(g).

Graphs.edges — Method

edges(g::EasyMultiGraph)

Return the edges of g as a list of ordered pairs (u,v) with u ≤ v. Edges are repeated in the list according to their multiplicity.

Graphs.ne — Method

ne(M::Matroid)

Return the number of elements in the matroid M.

LinearAlgebra.rank — Method

rank(M::Matroid, S::Set{T})::Int where T<:Integer
rank(M::Matroid)::Int

Return the rank of the set S in the matroid M.

We support alternative ways to invoke rank: For example, rank(M,a,b,c) and rank(M,[a,b,c]) are the same as rank(M,Set([a,b,c])).

Without S, return the rank of the matroid M.

Matroids._column_labels — Method

_column_labels(A::AbstractMatrix)

Create a dictionary mapping integers i to column vectors A[:,i].

Matroids._column_picker — Method

_column_picker(A::AbstractMatrix, S::Set{T}) where {T<:Integer}

Return the submatrix of A using the columns specified in S.

Matroids._default_labels — Method

_default_labels(m::Int)

Create a dictionary with keys 1:m and the value for x is x.

Matroids._random_set — Function

_random_set(n::Int, p::Real=0.5)

Create a random subset of {1,2,...,n} where each element is present in the set with probability p.

Matroids._set_check — Method

_set_check(S::Set{T}, m::Int) where {T<:Integer}

Check S is a subset of {1,2,...,m}. If not, throw an error.

Matroids._sort_order — Method

_sort_order(wt::Dict{T,R}) where {T<:Integer,R<:Real}

Return the keys of wt as a list such that their corresponding values are in nondecreasing order.

Matroids.add! — Method

add!(g::EasyMultiGraph, u, v)

Add an edge to the multigraph by increasing its multiplicity by 1. Return true if successful.

Matroids.all_bases — Method

all_bases(M::Matroid)

Return an iterator that generates all the bases of M.

Matroids.basis — Method

basis(M::Matroid)::Set{Int}

Return a basis (maximum size independent set) of M.

Matroids.closure — Method

closure(M::Matroid, X::Set{T}) where {T<:Integer}

Compute the closure [a.k.a. span] of a set of elements X of a matroid.

Matroids.delete — Method

delete(M::Matroid, X)

Delete the elements in X from the matroid M. Here, X may be a Set or a Vector of integer values, or a single integer.

Matroids.disjoint_union — Method

disjoint_union(M1::Matroid, M2::Matroid)::Matroid

Form the disjoint union of two matroids. May be invoked as M1 + M2.

Matroids.dual — Method

dual(M::Matroid)

Create a new matroid that is the dual of M.

Matroids.find_label — Method

find_label(M::Matroid, lab)

Find an element of M whose label is lab, or return 0 if no such label exists.

Matroids.fuzzy_equal — Function

fuzzy_equal(M1::Matroid, M2::Matroid, reps::Int=1000, p::Real=0.5)::Bool

This is a randomized test to see if two matroids are equal. If the result is false, they are definitely not equal. If the result is true, they probably are.

Matroids.get_label — Method

get_label(M::Matroid, x::T) where {T<:Integer}

Return the label associated with element x of the matroid M.

Matroids.isflat — Method

isflat(M::Matroid, X::Set{T}) where {T<:Integer}

Return true is X is equal to its closure in the matroid M.

Matroids.isindependent — Method

isindependent(M::Matroid, S::Set{T})::Bool where {T<:Integer}

Check if the set S is independent in the matroid M.

Matroids.isloop — Method

isloop(M, x::T) where {T<:Integer}

Check if x is a loop in the matroid M.

Matroids.min_weight_basis — Method

min_weight_basis(M::Matroid, wt::Dict{T,R})::Set{Int} where {T<:Integer,R<:Real}

Return a minimum weight basis of M where the weights of the elements are specified by wt.

Matroids.random_basis — Method

random_basis(M::Matroid)::Set{Int}

Generate a random basis for M by assigning random weights to the elements of M and returning a minimum weight basis.

Matroids.rem! — Method

rem!(g::EasyMultiGraph, u, v)

Remove an edge from the multigraph by decreasing its multiplicity by 1. Return true if successful.

Matroids.reset_labels! — Method

reset_labels!(M::Matroid, labs::Dict)

Reset all labels for M to be labs.

Matroids.reset_labels! — Method

reset_labels!(M::Matroid)

Set the labels of the elements of M to their default, i.e., element x has label x.

Matroids.set_label! — Method

set_label!(M::Matroid, x::T, lab) where {T<:Integer}

Set the label of element x in matroid M to lab.