Library

Module

Matroids.Matroids — Module

Matroids

A Julia package for matroids.

The basic constructor is Matroid.

Matroids.AbstractMatroid

Base.size

Matroids.Matroid

Matroids.bases

Matroids.basis

Matroids.circuit

Matroids.circuit

Matroids.circuitclosures

Matroids.circuits

Matroids.closure

Matroids.cobasis

Matroids.cocircuit

Matroids.cocircuit

Matroids.cocircuits

Matroids.coclosure

Matroids.coloops

Matroids.corank

Matroids.corank

Matroids.flats

Matroids.fullcorank

Matroids.fullrank

Matroids.fundamentalcircuit

Matroids.fundamentalcocircuit

Matroids.groundset

Matroids.independentsets

Matroids.isbasis

Matroids.iscircuit

Matroids.isclosed

Matroids.iscobasis

Matroids.iscocircuit

Matroids.iscoclosed

Matroids.iscodependent

Matroids.iscoindependent

Matroids.isdependent

Matroids.isindependent

Matroids.isisomorphic

Matroids.isvalidmatroid

Matroids.loops

Matroids.maxcoindependent

Matroids.maxindependent

Matroids.nonbases

Matroids.rank

Matroids.rank

Types

Matroids.AbstractMatroid — Type

AbstractMatroid

An abstract type representing a matroid.

Functions

Matroids.Matroid — Function

Matroid(groundset=nothing, data=nothing; kwargs...)

Construct a matroid.

You will need to provide a graph or a matrix as data, or one of the optional kwargs.

Arguments

groundset : labels for the groundset of the matroid

data : a matrix or a graph

bases : optional; the set of bases of the matroid

independentsets : optional; the set of the independent sets of the matroid

circuits : optional; the set of circuits of the matroid

matroid : optional; an existing matroid

Base.size — Method

size(M)

Return the size of the groundset of the matroid M.

Matroids.bases — Method

bases(M)

Return all the bases of the matroid M.

Matroids.basis — Method

basis(M)

Return an arbitrary basis of the matroid M.

Matroids.circuit — Method

circuit(M, X)

Return a circuit of M contained in X. Otherwise raise an error. X should be a subset of the groundset of M.

Matroids.circuit — Method

circuit(M)

Return a circuit of M if one exists. Otherwise raise an error.

Matroids.circuitclosures — Method

circuitclosures(M)

Return all the circuit closures of the matroid M.

Matroids.circuits — Method

circuits(M)

Return all the circuits of the matroid M.

Matroids.closure — Method

closure(M, X)

Return the closure of X in matroid M. X should be a subset of the groundset of M.

Matroids.cobasis — Method

cobasis(M)

Return an arbitrary cobasis of the matroid M.

Matroids.cocircuit — Method

cocircuit(M, X)

Return a cocircuit of M contained in X. Otherwise raise an error. X should be a subset of the groundset of M.

Matroids.cocircuit — Method

cocircuit(M)

Return a cocircuit of M if one exists. Otherwise raise an error.

Matroids.cocircuits — Method

cocircuits(M)

Return all the cocircuits of the matroid M.

Matroids.coclosure — Method

coclosure(M, X)

Return the coclosure of X in matroid M. X should be a subset of the groundset of M.

Matroids.coloops — Method

coloops(M)

Return all the coloops of the matroid M.

Matroids.corank — Method

corank(M, X)

Return the corank of the subset X in the matroid M. X shold be a subset of the groundset of M.

Matroids.corank — Method

corank(M)

Return the corank of the matroid M. Equivalent to fullcorank(M).

Matroids.flats — Method

flats(M, r)

Return all the flats of rank r in the matroid M.

Matroids.fullcorank — Method

fullcorank(M)

Return the corank of the matroid M. Equivalent to corank(M).

Matroids.fullrank — Method

fullrank(M)

Return the rank of the matroid M. Equivalent to rank(M).

Matroids.fundamentalcircuit — Method

fundamentalcircuit(M, B, e)

Return the B-fundamental circuit using e in matroid M. This is the unique matroid circuit contained in $B \cup e$. B should be a basis of M and e should be an element of M not in B.

Matroids.fundamentalcocircuit — Method

fundamentalcircuit(M, B, e)

Return the B-fundamental cocircuit using e in matroid M. This is the unique matroid cocircuit that intersects B only at e. B should be a basis of M and e should be an element of M not in B.

Matroids.groundset — Method

groundset(M)

Return the groundset of the matroid M.

Matroids.independentsets — Method

independentsets(M)

Return all the independent sets of the matroid M.

Matroids.isbasis — Method

isbasis(M, X)

Test whether the subset X is a basis in the matroid M. X should be a subset of the groundset of M.

Matroids.iscircuit — Method

iscircuit(M, X)

Test whether the subset X is a circuit in the matroid M. X should be a subset of the groundset of M.

Matroids.isclosed — Method

isclosed(M, X)

Test whether the subset X is closed in the matroid M. X should be a subset of the groundset of M.

Matroids.iscobasis — Method

iscobasis(M, X)

Test whether the subset X is a cobasis in the matroid M. X should be a subset of the groundset of M.

Matroids.iscocircuit — Method

iscocircuit(M, X)

Test whether the subset X is a cocircuit in the matroid M. X should be a subset of the groundset of M.

Matroids.iscoclosed — Method

iscoclosed(M, X)

Test whether the subset X is coclosed in the matroid M. X should be a subset of the groundset of M.

Matroids.iscodependent — Method

iscodependent(M, X)

Test whether the subset X is codependent in the matroid M. X should be a subset of the groundset of M.

Matroids.iscoindependent — Method

iscoindependent(M, X)

Test whether the subset X is coindependent in the matroid M. X should be a subset of the groundset of M.

Matroids.isdependent — Method

isdependent(M, X)

Test whether the subset X is dependent in the matroid M. X should be a subset of the groundset of M.

Matroids.isindependent — Method

isindependent(M, X)

Test whether the subset X is independent in the matroid M. X should be a subset of the groundset of M.

Matroids.isvalidmatroid — Method

isvalidmatroid(M)

Test whether M is a valid matroid, i.e. it satisfies the matroid axioms.

Matroids.loops — Method

loops(M)

Return all the loops of the matroid M.

Matroids.maxcoindependent — Method

maxcoindependent(M, X)

Return a maximal coindependent subset of X in matroid M. X should be a subset of the groundset of M.

Matroids.maxindependent — Method

maxindependent(M, X)

Return a maximal independent subset of X in matroid M. X should be a subset of the groundset of M.

Matroids.nonbases — Method

nonbases(M)

Return all the nonbases of the matroid M.

Matroids.rank — Method

rank(M, X)

Return the rank of the subset X in the matroid M. X shold be a subset of the groundset of M.

Matroids.rank — Method

rank(M)

Return the rank of the matroid M. Equivalent to fullrank(M).

Matroids.isisomorphic — Method

isisomorphic(M, N)

Test whether matroids M and N are isomorphic.